remove hardcoded items from the list and add new functionality to add items in a separate file
